Add files via upload
This is a customer segmentation that me and my team collaborated on. It was done for an eCommerce business in order to identify their key customers, improve the customer retention and the customer conversion rate.

Feel free to ask any questions about the data analysis, the reason behind choosing those segments or anything related to the presentation.