Add files via upload
The eCommerce company called Prism were looking to change their current subscription tier. This is the new version that me and my team worked on. The presentation contains various reasons on why the current subscription model is not working and how different alternatives could create a better relationship with our customer base.